(refactor) e2e: split feed-and-posts tests by feature, fix search-posts pagination
Split the monolithic feed-and-posts.e2e.test.ts into 6 per-feature files
(get-feed, get-post, get-post-engagers, get-post-stats, search-posts,
get-profile-activity) so individual features can be tested independently.

Drop URN reconstruction from search-posts URLs — URNs are not extractable
from search results. Switch search-posts cursor pagination from broken
URN-based to index-based (number type). Add E2E pagination test.

Refactor get-profile-activity to discover profile public ID from feed
post author URLs instead of requiring LHREMOTE_E2E_PROFILE_PUBLIC_ID.
